refactor: shrink ts cli dispatcher
